+++ src/main.rs
@@ -8,8 +8,6 @@ fn compile_from_cli() -> Result<()> {
     let runtime = Runtime::from_cli_args()?;
     let template = Template::compile(&runtime)?;
 
-    template.write_object_file(&runtime)?;
-
     let target = runtime.target()?;
     if runtime.options.backup {
         target.write_backup()?;