+++ docs/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Features
 
+* Using object code files (.csvpo) during compilation
 * Code coverage reporting
 
 ### **Breaking (Language) Changes**
@@ -9,6 +10,8 @@
 * All builtin variables & functions are removed.  All of the functionality provided by them could be done with native spreadsheet functions anyway.  And we're adding support for module loading which means they could just be implemented as a lib instead.
 * Rename `![[expand]]` to `![[fill]]` and all related code references
 * Rename `[[format]]`  to `[[text]]` 
+* Rename `Error::ObjectWriteError` to `Error::ObjectCodeError`
+* `Template.write_object_file` made crate-private
 
 ## v0.4.1